removed unused files, reorganized components files
