+++ src/components/Editor/BottomEditorPanel/index.tsx
