+++ src/components/Editor/CadenceEditor/ControlPanel/Arguments/styles.tsx
@@ -1,43 +1,31 @@
-import LegacyButton from 'components/LegacyButton';
-import { motion } from 'framer-motion';
-import React from 'react';
 import styled from 'styled-components';
-import theme from '../../../theme';
-
-export const MotionBox = (props: any) => {
-  const { children, dragConstraints } = props;
-  return (
-    <motion.div
-      className="drag-box"
-      drag={true}
-      dragConstraints={dragConstraints}
-      dragElastic={0}
-      whileDrag={{ scale: '1.1', opacity: '0.5' }}
-      dragMomentum={false}
-    >
-      {children}
-    </motion.div>
-  );
-};
+import theme from '../../../../../theme';
 
 interface HoverPanelProps {
-  width?: string;
+  minWidth?: string;
 }
 
 export const HoverPanel = styled.div<HoverPanelProps>`
-  min-width: 362px;
-  max-width: 500px;
-  padding: 20px;
-  border-radius: 4px;
+  min-width: ${({ minWidth }) => minWidth};
+  max-width: 362px;
+  border-radius: 8px;
   background-color: #fff;
-  box-shadow: 10px 10px 20px #c9c9c9, -10px -10px 20px #ffffff;
+  border: 1px solid #abb3bf;
+  box-shadow: 0px 4px 40px rgba(0, 0, 0, 0.08);
+`;
+
+interface HidableProps {
+  hidden: Boolean;
+}
+export const Hidable = styled.div<HidableProps>`
+  display: ${({ hidden = false }) => (hidden ? 'none' : 'block')};
 `;
 
 export const Heading = styled.div`
   display: flex;
   align-items: center;
   justify-content: space-between;
-  margin-bottom: 16px;
+  padding: 0.5rem;
 `;
 
 interface TitleProps {
@@ -70,6 +58,7 @@ export const Controls = styled.div`
   display: flex;
   justify-content: flex-end;
   align-items: center;
+  width: 100%;
   cursor: pointer;
 `;
 
@@ -105,27 +94,35 @@ export const List = styled.div<ListProps>`
   display: ${({ hidden }) => (hidden ? 'none' : 'grid')};
   grid-gap: 12px;
   grid-template-columns: 100%;
-  margin-bottom: 24px;
   max-height: 350px;
   overflow-y: auto;
+  padding: 0.5rem;
 `;
 
-export const SignersContainer = styled.div`
-  margin-bottom: 20px;
-`;
+export const SignersContainer = styled.div``;
 
 interface ControlContainerProps {
   isOk: boolean;
   progress: boolean;
+  showPrompt?: boolean;
 }
+interface MessageProps {
+  isOk?: boolean;
+}
+
 export const ControlContainer = styled.div<ControlContainerProps>`
-  display: flex;
+  display: ${({ showPrompt }) => (showPrompt ? 'block' : 'flex')};
   align-items: center;
   justify-content: space-between;
-  color: ${({ isOk, progress }) => {
+  padding: 8px;
+  width: 100%;
+  border-top: 1px solid #abb3bf;
+  color: ${({ isOk, progress, showPrompt }) => {
     switch (true) {
       case progress:
         return '#a2a2a2';
+      case isOk && showPrompt:
+        return theme.colors.error;
       case isOk:
         return '#2bb169';
       default:
@@ -145,7 +142,7 @@ export const ToastContainer = styled.div`
   color: ${theme.colors.darkPrimary};
 `;
 
-export const StatusMessage = styled.div`
+export const StatusMessage = styled.div<MessageProps>`
   @keyframes spin {
     from {
       transform: rotate(0deg);
@@ -156,6 +153,8 @@ export const StatusMessage = styled.div`
   }
 
   display: flex;
+  width: ${({ isOk }) => (isOk ? '100%' : '50%')};
+  padding: ${({ isOk }) => (isOk ? '1rem' : 'unset')};
   justify-content: flex-start;
   font-size: 16px;
   svg {
@@ -172,6 +171,8 @@ export const ErrorsContainer = styled.div`
   grid-gap: 10px;
   grid-template-columns: 100%;
   margin-bottom: 12px;
+  background: #f6f7f9;
+  border-radius: 8px 8px 8px 0px;
 `;
 
 export const SingleError = styled.div`
@@ -180,8 +181,10 @@ export const SingleError = styled.div`
   align-items: baseline;
   box-sizing: border-box;
   padding: 10px;
-  border-radius: 4px;
   font-size: 14px;
+  background: #ffffff;
+  box-shadow: 0px 4px 40px rgba(0, 0, 0, 0.08);
+  border-radius: 8px;
   &:hover {
     background-color: rgba(244, 57, 64, 0.15);
 
@@ -196,15 +199,9 @@ export const SingleError = styled.div`
 `;
 
 export const ErrorIndex = styled.div`
-  width: 20px;
-  height: 20px;
-  background-color: rgba(0, 0, 0, 0.15);
   display: flex;
   align-items: center;
-  justify-content: center;
-  border-radius: 20px;
-  margin-right: 8px;
-  flex: 0 0 auto;
+  justify-content: space-between;
 `;
 
 export const ErrorMessage = styled.p`
@@ -232,49 +229,4 @@ export const SignersError = styled.p`
   svg {
     margin-right: 0.5em;
   }
-`;
-
-export const Confirm = styled(LegacyButton)`
-  background-color: ${theme.colors.error};
-  color: #fff;
-  margin-right: 0.5rem;
-
-  &:active {
-    background-color: #cf3529;
-  }
-`;
-
-export const Cancel = styled(LegacyButton)`
-  background-color: ${theme.colors.background};
-  color: ${theme.colors.text};
-  border: 1px solid ${theme.colors.greyBorder};
-
-  &:active {
-    background-color: #dedede;
-  }
-`;
-
-export const PromptActionsContainer = styled.div`
-  display: flex;
-  justify-content: flex-start;
-`;
-
-interface StatusIconProps {
-  isOk: boolean;
-  progress: boolean;
-  showPrompt: boolean;
-}
-export const StatusIcon = styled.div<StatusIconProps>`
-  color: ${({ isOk, progress, showPrompt }) => {
-    switch (true) {
-      case progress:
-        return '#a2a2a2';
-      case isOk && showPrompt:
-        return theme.colors.warning;
-      case isOk:
-        return '#2bb169';
-      default:
-        return '#EE431E';
-    }
-  }};
 `;
