+++ src/components/Editor/EditorPanels.tsx
@@ -4,18 +4,18 @@ import 'allotment/dist/style.css';
 import { EntityType } from 'providers/Project';
 import { useProject } from 'providers/Project/projectHooks';
 import React, { useState } from 'react';
-import theme from '../theme';
+import theme from '../../theme';
 import { SXStyles } from 'src/types';
 import { Box, Flex } from 'theme-ui';
 import { storageMapByIndex } from 'util/accounts';
 import BottomEditorPanel from './BottomEditorPanel';
-import BottomEditorPanelHeader from './BottomEditorPanelHeader';
-import Button from './Button';
+import BottomEditorPanelHeader from './BottomEditorPanel/BottomEditorPanelHeader';
+import Button from '../Button';
 import CadenceEditor from './CadenceEditor';
-import CopyIcon from './Icons/CopyIcon';
-import ExplorerTransactionIcon from './Icons/ExplorerTransactionIcon';
-import ExplorerScriptIcon from './Icons/ExplorerScriptIcon';
-import ExplorerContractIcon from './Icons/ExplorerContractIcon';
+import CopyIcon from '../Icons/CopyIcon';
+import ExplorerTransactionIcon from '../Icons/ExplorerTransactionIcon';
+import ExplorerScriptIcon from '../Icons/ExplorerScriptIcon';
+import ExplorerContractIcon from '../Icons/ExplorerContractIcon';
 import useClipboard from 'react-use-clipboard';
 import { FaClipboardCheck } from 'react-icons/fa';
